Updated for FAST-324 - Create users in SFCC during order creation, FAST-347 - Not able to Start Fast checkout
