+++ sfcc_cartridge/int_fast_sfra/cartridge/scripts/hooks.json
@@ -9,6 +9,10 @@
             "script": "./basket/basketHookScripts.js"
 		},
 		{
+			"name": "dw.ocapi.shop.basket.afterPATCH",
+            "script": "./basket/basketHookScripts.js"
+		},
+		{
 			"name": "dw.ocapi.shop.order.afterPOST",
 			"script": "./order/orderHookScripts.js"
 		},
